# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "嘉阅圈私域社群平台后端服务，提供内容社区、互动打卡、小组交流、私信通知、知识课程等核心能力。\n统一响应格式：{\"success\":bool,\"data\":any,\"message\":string,\"error\":{\"code\":string,\"message\":string}}",
        "title": "嘉阅圈后端 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3100",
    "basePath": "/api",
    "paths": {
...
        "/circles": {
            "get": {
                "description": "公开接口。支持按关键词搜索与按拥有者过滤，返回分页的圈子列表。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "圈子"
                ],
                "summary": "分页查询圈子列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "名称或描述关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "拥有者用户 ID",
                        "name": "ownerId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认 20",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "需登录且需平台管理员权限。创建圈子后自动将当前管理员以 OWNER 角色加入并初始化成员计数为 1。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "圈子"
                ],
                "summary": "创建圈子",
                "parameters": [
                    {
                        "description": "创建圈子请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数格式错误或名称不合规",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "403": {
                        "description": "仅平台管理员可创建圈子",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/circles/mine": {
            "get": {
                "description": "需登录。返回当前登录用户已加入的圈子列表。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "圈子"
                ],
                "summary": "获取当前用户已加入的圈子",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/circles/{id}": {
            "get": {
                "description": "公开接口，可选登录。返回圈子详情，非成员或非管理员时隐藏成员名单。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "圈子"
                ],
                "summary": "获取圈子详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "圈子 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "无效的圈子 ID",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "圈子不存在",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "需登录。仅圈子拥有者或平台管理员可更新圈子信息，更新成功后返回最新详情。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "圈子"
                ],
                "summary": "更新圈子",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "圈子 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新圈子请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数格式错误或无效的圈子 ID",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "403": {
                        "description": "只有圈子拥有者或管理员才能更新圈子信息",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "圈子不存在",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "需登录。仅圈子拥有者或平台管理员可删除圈子，删除成功后返回 deleted=true。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "圈子"
                ],
                "summary": "删除圈子",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "圈子 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "无效的圈子 ID",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "403": {
                        "description": "只有圈子拥有者或管理员才能删除圈子",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "圈子不存在",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/circles/{id}/join": {
            "post": {
                "description": "需登录。当前用户以 MEMBER 角色加入指定圈子，已加入则返回 409，加入成功后成员计数 +1。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "圈子"
                ],
                "summary": "加入圈子",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "圈子 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "无效的圈子 ID",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "圈子不存在",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "409": {
                        "description": "已经是该圈子的成员",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/circles/{id}/leave": {
            "post": {
                "description": "需登录。当前用户退出指定圈子，退出后成员计数 -1；圈子拥有者不可退出，需先转让圈子。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "圈子"
                ],
                "summary": "退出圈子",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "圈子 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "无效的圈子 ID 或圈子拥有者不能退出",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "圈子不存在",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "409": {
                        "description": "您不是该圈子的成员",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "yjedu-reading-club-server_internal_pkg_response.ApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "yjedu-reading-club-server_internal_pkg_response.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiError"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 \"Bearer \u003ctoken\u003e\"，token 通过 /api/auth/login/phone 或 /api/auth/weapp/login 获取。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}